PP-9955 refactor `demo-payment` folder structure
- refactor `demo-payment` folder structure to follow GDS' project directory structure
