+++ app/demo-payment/failed.njk
@@ -1,4 +1,4 @@
-{% extends "../layout.njk" %}
+{% extends "../views/layout.njk" %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 
 {% block pageTitle %}After a failed payment - GOV.UK Pay{% endblock %}
