+++ app/demo-payment/payment-success.controller.js
@@ -2,10 +2,10 @@
 
 const urlJoin = require('url-join')
 
-const staticify = require('../../../server').staticify
-const { SELFSERVICE_TRANSACTIONS_URL, DOCS_URL } = require('../../../config/index')
+const staticify = require('../../server').staticify
+const { SELFSERVICE_TRANSACTIONS_URL, DOCS_URL } = require('../../config')
 
-const CONFIRM_SUCCESS_VIEW = 'confirm-demo-payment/success'
+const CONFIRM_SUCCESS_VIEW = './success'
 
 const data = {
   scenarioDocsLink: urlJoin(DOCS_URL, '/#confirmation-page'),
