+++ app/routes.js
@@ -11,8 +11,8 @@ const healthcheckCtrl = require('./controllers/healthcheck.controller')
 const friendlyUrlRedirectCtrl = require('./controllers/friendly-url-redirect.controller')
 const prePaymentCtrl = require('./controllers/pre-payment.controller')
 const completeCtrl = require('./controllers/payment-complete.controller')
-const failedCtrl = require('./controllers/demo-payment/payment-failed.controller')
-const successCtrl = require('./controllers/demo-payment/payment-success.controller')
+const failedCtrl = require('./demo-payment/payment-failed.controller')
+const successCtrl = require('./demo-payment/payment-success.controller')
 const amountCtrl = require('./payment-links/amount/amount.controller')
 const referenceCtrl = require('./payment-links/reference/reference.controller')
 const referenceConfirmCtrl = require('./payment-links/reference-confirm/reference-confirm.controller')
